requirements.txt
asgiref
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
mysqlclient
requests
sqlparse
redis
dj_database_url
celery
# Sécurité
django-environ
python-dotenv
gunicorn
# Auth & Permissions
djangorestframework-simplejwt
django-cors-headers

# Gestion Cache & Performance 
django-redis
django-cacheops
whitenoise